pyproject.toml
[project]
name = "review-intelligence-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "openai>=2.33.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
